Created Queen Move Generator
In the queen move generator, I realized I could reuse the code of rook and bishop move generator since a queen = bishop + rook.

I moved the logic of bishop and rook move generators to separate files, and changed the function so it could be used for queen move generator, and rook and bishop move generation.

Then I called the functions as needed  in the generator files.